requirements.txt
requests
python-dotenv
Pillow
openpyxl
